refactor: split workflow controller into separate files
